+++ client/src/features/place/AboutPanel.tsx
@@ -37,7 +37,7 @@ const AboutPanel: React.FC<IProps> = ({ entity }) => {
         {types.length > 0 && (
           <LinkContainer
             content={types}
-            label="Types"
+            label="Categorized As"
             expandColumns
             itemSpacing="single"
             id="place-types-link-container"